Finalize the Web Metric Data
