+++ src/components/home/Details.js
@@ -30,7 +30,6 @@ const Details = () => {
 
   const { All } = country;
   const list = Object.entries(country).slice();
-  // console.log(list);
 
   return (
     <section>
