+++ src/index.js
@@ -4,7 +4,6 @@ import { Provider } from 'react-redux';
 import './index.css';
 import App from './App';
 import store from './redux/configureStore';
-// import './App.css';
 
 ReactDOM.render(
   <React.StrictMode>
