+++ src/test-utils.js
@@ -1,6 +1,4 @@
-/* eslint-disable react/prop-types */
 import React from 'react';
-// eslint-disable-next-line import/no-extraneous-dependencies
 import { render as rtlRender } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { createStore, combineReducers, applyMiddleware } from 'redux';
@@ -24,12 +22,10 @@ function render(
     ...renderOptions
   } = {},
 ) {
-  function Wrapper({ children }) {
+  function Wrapper(children) {
     return <Provider store={store}>{children}</Provider>;
   }
   return rtlRender(ui, { wrapper: Wrapper, ...renderOptions });
 }
 
-// eslint-disable-next-line import/no-extraneous-dependencies
-export * from '@testing-library/react';
-export { render };
+export default render;
